Basic Chatbot Integration/requirements.txt
langchain_core
langchain_community
streamlit
python-dotenv
langserve
fastapi
uvicorn
sse_starlette
faiss-cpu
